working sunrise/sunset dimming function. fixed tooltips and button placement
